fix: Bet value의 type을 int로 수정

+++ src/main/java/blackjack/domain/Bet.java
@@ -2,20 +2,20 @@ package blackjack.domain;
 
 public class Bet {
 
-    private final double value;
+    private final int value;
 
-    public Bet(final double value) {
+    public Bet(final int value) {
         validateBettingMoney(value);
         this.value = value;
     }
 
-    private void validateBettingMoney(final double value) {
+    private void validateBettingMoney(final int value) {
         if (value <= 0) {
             throw new IllegalArgumentException("베팅 금액은 양수여야 합니다.");
         }
     }
 
-    public double getValue() {
+    public int getValue() {
         return this.value;
     }
 }
